requirements.txt
fastapi==0.104.1
pydantic[email]==2.4.2
apscheduler==3.10.4
sqlalchemy==2.0.23
uvicorn==0.24.0
asyncpg==0.29.0
bcrypt==4.0.1
cryptography==41.0.5
pyjwt==2.8.0
greenlet==3.0.1
requests==2.31.0

